# api/openapi.yaml
openapi: 3.1.0
info:
  title: EVE-Cyno Data API
  summary: Deterministic EVE Online data, fit calculations and tools. No language model in the loop.
  description: |-
    Deterministic EVE Online data and calculations: community-fit search, fit breakdown and statistics (the Gofa dogma engine), SDE item search, and the tool layer behind the EVE-Cyno assistant, also reachable over MCP. Nothing here calls a language model; the same inputs give the same answers.

    ## Conventions

    - Errors are JSON, always: `{"error":{"code":"...","message":"..."}}`. The codes are stable (see `ErrorDetail`).
    - Every response carries `X-Request-ID`: your value when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one.
    - Requests are rate limited per caller (a sliding window; the budget is stated on each operation): per API key when one is sent, otherwise per client IP. A refused request is `429` with `Retry-After`.
    - Request bodies are capped (256 KiB on the fit routes, 64 KiB on tools and MCP); an oversize body is `413`.
...

    ## Attribution and licences

    © 2014 Fenris Creations hf. All rights reserved. 'EVE', 'EVE Online', 'Fenris Creations', and all related logos and images are trademarks or registered trademarks of Fenris Creations hf. This material is used with limited permission of Fenris Creations. No official affiliation or endorsement by Fenris Creations is stated or implied. Created under the EVE Developer License Agreement.

...
servers:
...
paths:
  /fits/search:
    get:
      tags:
        - fits
      summary: Search community fits
      description: |-
        Searches the community-fit corpus (EVE Workbench, Abysstracker and other sources named in each hit). The facet parameters filter the corpus; with `q` a filtered vector search ranks the hits. When the facets match nothing they are relaxed one at a time, last first: `relaxed` is then true and `dropped` names what was removed. `ship` may be a loose hull name (`megath` finds Megathron). No language model is involved. Every hit carries an `attribution` object (`source`, `source_url`, `author` when known, `license`, `credit` and `license_url` when the terms require them); show it wherever you show the fit. The fits are community data served ranked and capped, not for bulk export.

        Rate limit: 120 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: searchFits
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: ship
          in: query
          required: false
          description: Hull name; a partial or loose name is resolved against the SDE.
          schema:
            type: string
        - name: activity
          in: query
          required: false
          description: Activity tag, e.g. `pve` or `pvp`.
          schema:
            type: string
        - name: tag
          in: query
          required: false
          description: Extra tag, e.g. `abyss` or `ratting`.
          schema:
            type: string
        - name: filament
          in: query
          required: false
          description: Abyssal filament type.
          schema:
            type: string
        - name: cost
          in: query
          required: false
          description: Cost class tag, e.g. `cheap`.
          schema:
            type: string
        - name: source
          in: query
          required: false
          description: Restrict to one corpus source; empty searches the default source set.
          schema:
            type: string
        - name: clone
          in: query
          required: false
          description: '`alpha` keeps only alpha-clone fits, `omega` only omega-only fits; empty keeps both.'
          schema:
            type: string
            enum:
              - alpha
              - omega
        - name: q
          in: query
          required: false
          description: Free-text query; empty lists by the facets alone.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: 'Maximum hits, clamped to 1-24 (the response `limit` and `note` say so). There is no offset, page or cursor: one query reaches one ranked page, and a request carrying such a parameter is refused with `invalid_request`.'
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: The ranked fits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitSearchResult'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "429":
          $ref: '#/components/responses/RateLimited'
        "502":
          $ref: '#/components/responses/UpstreamError'
        "503":
          $ref: '#/components/responses/Unavailable'
...
  /fit/stats:
    post:
      tags:
        - fits
      summary: Compute fit statistics
      description: |-
        Parses an EFT block and returns the Gofa engine's DPS, tank, capacitor, navigation, targeting and drone numbers. Deterministic; the same engine produces the numbers of the `compute_fit_stats` tool.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: fitStats
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitStatsRequest'
      responses:
        "200":
          description: The computed statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FitStats'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "502":
          $ref: '#/components/responses/UpstreamError'
        "503":
          $ref: '#/components/responses/Unavailable'
  /fit/suggest:
    post:
      tags:
        - fits
      summary: Suggest modules for an empty slot
      description: |-
        Ranks the modules that can fill a slot tier of the given fit: how often similar community fits carry them (when the corpus is up), their meta tier, and whether they fit the remaining CPU and powergrid. Without the corpus the ranking falls back to the SDE alone.

        Rate limit: 60 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: fitSuggest
      parameters:
        - $ref: '#/components/parameters/RequestID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FitSuggestRequest'
      responses:
        "200":
          description: At most 12 ranked suggestions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Suggestion'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "503":
          $ref: '#/components/responses/Unavailable'
...
  /tool/appraise_items:
    post:
      tags:
        - tools
      summary: Appraise a list of EVE Online items using Janice.
      description: |-
        Appraise a list of EVE Online items using Janice. Returns total buy/sell/split values in ISK and a link to the full appraisal.

        Sources: Janice.

        Access tier: **byo-key**. Runs only with your own Janice API key in the `X-Janice-Key` header (`400 janice_key_required` without it); the key is used for this request only and never logged or stored.

        Rate limit: 30 requests per minute per caller (per API key when one is sent, otherwise per client IP).
      operationId: appraise_items
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - name: X-Janice-Key
          in: header
          required: true
          description: Your own Janice API key (https://janice.e-351.com/). It is used for this request only, sent to Janice and nowhere else, and is never logged or stored by this service. This service never appraises with its own key.
          schema:
            type: string
            maxLength: 256
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: string
                  description: Items in multi-buy or EFT format, one per line with quantity, e.g. 'Tritanium x1000'
              required:
                - items
      responses:
        "200":
          description: The tool's answer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResponse'
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "413":
          $ref: '#/components/responses/PayloadTooLarge'
        "429":
          $ref: '#/components/responses/RateLimited'
        "500":
          $ref: '#/components/responses/InternalError'
        "503":
          $ref: '#/components/responses/Unavailable'
        "504":
          $ref: '#/components/responses/Timeout'
...
components:
  schemas:
...
    CapStats:
      type: object
      properties:
        Capacity:
          type: number
        RechargeSec:
          type: number
        Stable:
          type: boolean
        StablePct:
          type: number
        SecondsToEmpty:
          type: number
      required:
        - Capacity
        - RechargeSec
        - Stable
        - StablePct
        - SecondsToEmpty
...
    DPSStats:
      type: object
      properties:
        Theoretical:
          type: number
        Applied:
          type: number
        PerWeapon:
          type: array
          items:
            $ref: '#/components/schemas/WeaponDPS'
      required:
        - Theoretical
        - Applied
        - PerWeapon
...
    DroneStats:
      type: object
      properties:
        DPS:
          type: number
        BandwidthUsed:
          type: number
        BandwidthAvail:
          type: number
        BayUsed:
          type: number
        BayAvail:
          type: number
        InBay:
          type: integer
        Active:
          type: integer
        MaxActive:
          type: integer
        ActiveTypeIDs:
          type: array
          items:
            type: integer
      required:
        - DPS
        - BandwidthUsed
        - BandwidthAvail
        - BayUsed
        - BayAvail
        - InBay
        - Active
        - MaxActive
        - ActiveTypeIDs
    Error:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
          description: Every non-2xx response has exactly this shape.
      required:
        - error
...
    FitSearchHit:
      type: object
      properties:
        ship_name:
          type: string
        fit_name:
          type: string
        source:
          type: string
        source_url:
          type: string
        tags:
          type: array
          items:
            type: string
        views:
          type: integer
        score:
          type: number
        tested:
          type: boolean
        video:
          type: boolean
        alpha:
          type: boolean
        eft:
          type: string
        relevance:
          type: number
        attribution:
          $ref: '#/components/schemas/Attribution'
      required:
        - ship_name
        - fit_name
        - source
        - source_url
        - tags
        - views
        - score
        - tested
        - video
        - alpha
        - eft
        - relevance
        - attribution
    FitSearchResult:
      type: object
      properties:
        total:
          type: integer
        relaxed:
          type: boolean
        dropped:
          type: array
          items:
            type: string
        hits:
          type: array
          items:
            $ref: '#/components/schemas/FitSearchHit'
        ship_counts:
          type: object
          additionalProperties:
            type: integer
        limit:
          type: integer
        note:
          type: string
      required:
        - total
        - relaxed
        - dropped
        - hits
        - limit
...
    FitStats:
      type: object
      properties:
        DPS:
          $ref: '#/components/schemas/DPSStats'
        Volley:
          type: number
        Tank:
          $ref: '#/components/schemas/TankStats'
        Capacitor:
          $ref: '#/components/schemas/CapStats'
        Navigation:
          $ref: '#/components/schemas/NavStats'
        Range:
          type: array
          items:
            $ref: '#/components/schemas/RangeStats'
        Targeting:
          $ref: '#/components/schemas/TargetStats'
        Drones:
          $ref: '#/components/schemas/DroneStats'
        Estimated:
          type: boolean
        Unmodelled:
          type: array
          items:
            type: string
      required:
        - DPS
        - Volley
        - Tank
        - Capacitor
        - Navigation
        - Range
        - Targeting
        - Drones
        - Estimated
        - Unmodelled
    FitStatsRequest:
      type: object
      properties:
        eft:
          type: string
          description: The fit as an EFT block.
        activeDrones:
          type: array
          description: Type IDs of the drones to launch, in priority order; omitted means automatic.
          items:
            type: integer
      required:
        - eft
    FitSuggestRequest:
      type: object
      properties:
        eft:
          type: string
          description: The current fit as an EFT block.
        slot:
          type: string
          description: The slot tier to fill.
          enum:
            - high
            - mid
            - low
            - rig
      required:
        - eft
        - slot
...
    NavStats:
      type: object
      properties:
        MaxVelocity:
          type: number
        AlignTimeSec:
          type: number
        SignatureRadius:
          type: number
      required:
        - MaxVelocity
        - AlignTimeSec
        - SignatureRadius
    RangeStats:
      type: object
      properties:
        Weapon:
          type: string
        OptimalM:
          type: number
        FalloffM:
          type: number
      required:
        - Weapon
        - OptimalM
        - FalloffM
...
    Source:
      type: object
      properties:
        name:
          type: string
          description: The upstream's name.
        url:
          type: string
          description: Where the upstream lives.
        license:
          type: string
          description: The terms the data is used under.
      required:
        - name
        - url
        - license
    Suggestion:
      type: object
      properties:
        typeID:
          type: integer
        name:
          type: string
        slot:
          type: string
        metaTier:
          type: string
        cpu:
          type: number
        pg:
          type: number
        fits:
          type: boolean
        reason:
          type: string
      required:
        - typeID
        - name
        - slot
        - metaTier
        - cpu
        - pg
        - fits
        - reason
...
    TankStats:
      type: object
      properties:
        ShieldHP:
          type: number
        ArmorHP:
          type: number
        HullHP:
          type: number
        ShieldResists:
          $ref: '#/components/schemas/ResistProfile'
        ArmorResists:
          $ref: '#/components/schemas/ResistProfile'
        HullResists:
          $ref: '#/components/schemas/ResistProfile'
        TotalEHP:
          type: number
        ActiveRepPerSec:
          type: number
      required:
        - ShieldHP
        - ArmorHP
        - HullHP
        - ShieldResists
        - ArmorResists
        - HullResists
        - TotalEHP
        - ActiveRepPerSec
    TargetStats:
      type: object
      properties:
        LockRangeM:
          type: number
        ScanResolution:
          type: number
        MaxLockedTargets:
          type: integer
        SensorStrength:
          type: number
      required:
        - LockRangeM
        - ScanResolution
        - MaxLockedTargets
        - SensorStrength
    ToolResponse:
      type: object
      properties:
        tool:
          type: string
          description: The tool that ran.
        version:
          type: string
          description: The build version of the service.
        result:
          type: object
          properties:
            text:
              type: string
              description: The answer as the chat assistant reads it.
            data:
              type: "null"
              description: The typed answer, or null when the tool has none or the call ended in a text-only answer (an unknown id, a parse failure); `text` then says why.
          required:
            - text
            - data
        attribution:
          type: array
          description: The upstreams behind the answer; attribution is a condition of several of their licences.
          items:
            $ref: '#/components/schemas/Source'
      required:
        - tool
        - version
        - result
        - attribution
...
  responses:
    BadRequest:
      description: 'The request is malformed: invalid JSON, or a missing or invalid parameter or header (`invalid_request`, `janice_key_required`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
    InternalError:
      description: The operation failed inside the service (`internal_error`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
...
    PayloadTooLarge:
      description: The request body is over the size cap (`payload_too_large`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    RateLimited:
      description: The caller is over the rate limit; retry after the number of seconds in `Retry-After` (`rate_limited`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Timeout:
      description: The operation ran past its time budget (`timeout`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Unauthorized:
      description: The tool needs an API key and the request has none (`api_key_required`), or the key it sent is not valid (`invalid_api_key`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    Unavailable:
      description: A dependency this operation needs is not loaded, for example the SDE or the fit corpus (`unavailable`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
    UpstreamError:
      description: An upstream dependency failed (`upstream_error`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: An ID to correlate this request with the service's logs; echoed in the response.
      schema:
        type: string
        pattern: ^[A-Za-z0-9._-]{1,64}$
  headers:
    RequestID:
      description: 'The request''s ID: the caller''s `X-Request-ID` when it is a token of 1-64 characters from `[A-Za-z0-9._-]`, otherwise a generated one. Quote it when reporting a problem.'
      schema:
        type: string
    RetryAfter:
      description: Seconds until the caller's budget has room again.
      schema:
        type: integer
        minimum: 1
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      description: The same API key in the `X-API-Key` header, for clients that cannot set `Authorization`.
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      description: 'An API key as `Authorization: Bearer <key>`. Keyed tools need one; the other tools ignore it, but a key that is sent and not valid is refused (`401 invalid_api_key`).'
      scheme: bearer